+++ src/schemathesis/utils.py
@@ -255,7 +255,14 @@ GenericResponse = Union[requests.Response, WSGIResponse]  # pragma: no mutate
 def copy_response(response: GenericResponse) -> GenericResponse:
     """Create a copy of the given response as far as it makes sense."""
     if isinstance(response, requests.Response):
+        # Hooks are not copyable. Keep them out and copy the rest
+        hooks = None
+        if response.request is not None:
+            hooks = response.request.hooks["response"]
+            response.request.hooks["response"] = []
         copied_response = deepcopy(response)
+        if hooks is not None:
+            copied_response.request.hooks["response"] = hooks
         setattr(copied_response, "raw", response.raw)
         return copied_response
     # Can't deepcopy WSGI response due to generators inside (`response.freeze` doesn't completely help)